requirements.txt
fpdf2==2.8.2
matplotlib==3.10.8
numpy==2.4.2
pandas==3.0.0
Pillow==12.1.1
plotly==5.22.0
PyJWT==2.11.0
streamlit==1.50.0
